% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralcf-package.R
\docType{package}
\name{neuralcf-package}
\alias{neuralcf}
\alias{neuralcf-package}
\title{neuralcf: Neural Collaborative Filtering}
\description{
Keras layers and models for neural collaborative filtering.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/vgherard/neuralcf}
  \item Report bugs at \url{https://github.com/vgherard/neuralcf/issues}
}

}
\author{
\strong{Maintainer}: Valerio Gherardi \email{vgherard@sissa.it} (\href{https://orcid.org/0000-0002-8215-3013}{ORCID})

}
\keyword{internal}
